requirements.txt
cachelib==0.1.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cs50==6.0.4
cycler==0.10.0
Flask==1.1.2
Flask-Session==0.3.2
greenlet==1.0.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.4.0
numpy==1.20.2
Pillow==8.1.2
pyparsing==2.4.7
python-dateutil==2.8.1
requests==2.25.1
six==1.15.0
SQLAlchemy==1.4.3
sqlparse==0.4.1
termcolor==1.1.0
urllib3==1.26.4
Werkzeug==1.0.1
